requirements.txt
termcolor==1.1.0
Pillow==6.2.1
